For agent retries - include 500 server error and expand to include proof requests
