+++ server/vcr-server/agent_webhooks/views.py
@@ -4,9 +4,6 @@ import time
 import os
 import random
 
-import requests
-from requests.adapters import HTTPAdapter
-from requests.packages.urllib3.util.retry import Retry
 from django.conf import settings
 from drf_yasg.utils import swagger_auto_schema
 from rest_framework import permissions, status
@@ -14,7 +11,7 @@ from rest_framework.decorators import api_view, permission_classes
 from rest_framework.response import Response
 
 from api.v2.models.Credential import Credential as CredentialModel
-from api.v2.utils import log_timing_method, log_timing_event
+from api.v2.utils import log_timing_method, log_timing_event, call_agent_with_retry
 from agent_webhooks.utils.credential import Credential, CredentialManager
 from agent_webhooks.utils.issuer import IssuerManager
 
@@ -95,41 +92,6 @@ def agent_callback(request, topic):
 
 # create one global manager instance
 credential_manager = CredentialManager()
-
-def call_agent_with_retry(agent_url, post_method=True, payload=None, headers=None, retry_count=5, retry_wait=1):
-    """
-    Post with retry - if returned status is 503 unavailable retry a few times.
-    """
-    try:
-        session = requests.Session()
-        retry = Retry(
-            total=retry_count,
-            connect=retry_count,
-            status=retry_count,
-            status_forcelist=[502,503,504],
-            method_whitelist=['HEAD', 'TRACE', 'GET', 'POST', 'PUT', 'OPTIONS', 'DELETE'],
-            read=0,
-            redirect=0,
-            backoff_factor=retry_wait
-        )
-        adapter = HTTPAdapter(max_retries=retry)
-        session.mount('http://', adapter)
-        session.mount('https://', adapter)
-        if post_method:
-            resp = session.post(
-                agent_url,
-                json=payload,
-                headers=headers,
-            )
-        else:
-            resp = session.get(
-                agent_url,
-                headers=headers,
-            )
-        return resp
-    except Exception as e:
-        LOGGER.error("Agent connection raised exception, raise: " + str(e))
-        raise
 
 def handle_credentials(state, message):
     """
